+++ zcash_keys/src/keys.rs
@@ -37,6 +37,9 @@ use {
 #[cfg(feature = "orchard")]
 use orchard::{self, keys::Scope};
 
+#[cfg(all(feature = "sapling", feature = "unstable"))]
+use ::sapling::zip32::ExtendedFullViewingKey;
+
 #[cfg(feature = "sapling")]
 pub mod sapling {
     pub use sapling::zip32::{
@@ -111,6 +114,8 @@ impl Display for DerivationError {
     }
 }
 
+impl std::error::Error for DerivationError {}
+
 /// A version identifier for the encoding of unified spending keys.
 ///
 /// Each era corresponds to a range of block heights. During an era, the unified spending key
@@ -175,6 +180,8 @@ impl std::fmt::Display for DecodingError {
         }
     }
 }
+
+impl std::error::Error for DecodingError {}
 
 #[cfg(feature = "unstable")]
 impl Era {
@@ -674,6 +681,24 @@ impl UnifiedFullViewingKey {
             vec![],
         )
     }
+
+    #[cfg(all(feature = "sapling", feature = "unstable"))]
+    pub fn from_sapling_extended_full_viewing_key(
+        sapling: ExtendedFullViewingKey,
+    ) -> Result<UnifiedFullViewingKey, DerivationError> {
+        Self::from_checked_parts(
+            #[cfg(feature = "transparent-inputs")]
+            None,
+            #[cfg(feature = "sapling")]
+            Some(sapling.to_diversifiable_full_viewing_key()),
+            #[cfg(feature = "orchard")]
+            None,
+            // We don't currently allow constructing new UFVKs with unknown items, but we store
+            // this to allow parsing such UFVKs.
+            vec![],
+        )
+    }
+
     /// Construct a UFVK from its constituent parts, after verifying that UIVK derivation can
     /// succeed.
     fn from_checked_parts(