+++ zcash_keys/CHANGELOG.md
@@ -6,6 +6,12 @@ and this library adheres to Rust's notion of
 
 ## [Unreleased]
 
+### Added
+- `zcash_keys::encoding::decode_extfvk_with_network`
+- `impl std::error::Error for Bech32DecodeError`
+- `impl std::error::Error for DecodingError`
+- `impl std::error::Error for DerivationError`
+
 ## [0.3.0] - 2024-08-19
 ### Notable changes
 - `zcash_keys`: